# Fix (most) allocations