--- a/src/equations/ideal_mhd_multiion_1d.jl
+++ b/src/equations/ideal_mhd_multiion_1d.jl
@@ -114,15 +114,15 @@ function initial_condition_weak_blast_wave(x, t, equations::IdealMhdMultiIonEqua
   v1 = r > 0.5 ? 0.0 : 0.1882 * cos(phi)
   p = r > 0.5 ? 1.0 : 1.245
 
-  #prim = (0.01, 0.01, 0.01)
-  prim = (1.0, 1.0, 1.0)
-  for i in eachcomponent(equations)
-    prim = (prim..., 2^(i-1) * (1-2)/(1-2^ncomponents(equations)) * rho, v1, 0.0, 0.0, p)
-    #prim = (prim..., rho, v1, 0.0, 0.0, p)
-    #prim = (prim..., 1.0, 1.0, 0.0, 0.0, 100.0)
+  prim = zero(MVector{nvariables(equations), real(equations)})
+  prim[1] = 1.0
+  prim[2] = 1.0
+  prim[3] = 1.0
+  for k in eachcomponent(equations)
+    set_component!(prim, k, 2^(k-1) * (1-2)/(1-2^ncomponents(equations)) * rho, v1, 0, 0, p, equations)
   end
 
-  return prim2cons(SVector{nvariables(equations), real(equations)}(prim), equations)
+  return prim2cons(SVector(prim), equations)
 end
 
 # TODO: Add initial condition equilibrium
@@ -133,13 +133,12 @@ end
   
   v1_plus, v2_plus, v3_plus, vk1_plus, vk2_plus, vk3_plus = charge_averaged_velocities(u, equations)
 
-  f_B1 = 0.0
-  f_B2 = v1_plus * B2 - v2_plus * B1
-  f_B3 = v1_plus * B3 - v3_plus * B1
-
-  f = (f_B1, f_B2, f_B3)
-
   mag_en = 0.5*(B1^2 + B2^2 + B3^2)
+
+  f = zero(MVector{nvariables(equations), eltype(u)})
+  f[1] = 0.0
+  f[2] = v1_plus * B2 - v2_plus * B1
+  f[3] = v1_plus * B3 - v3_plus * B1
 
   for k in eachcomponent(equations)
     rho, rho_v1, rho_v2, rho_v3, rho_e = get_component(k, u, equations)
@@ -157,10 +156,10 @@ end
     f4 = rho_v1*v3 #- B1*B3
     f5 = (kin_en + gamma*p/(gamma - 1))*v1 + 2 * mag_en * vk1_plus[k] - B1*(vk1_plus[k] * B1 + vk2_plus[k] * B2 + vk3_plus[k] * B3)
 
-    f = (f..., f1, f2, f3, f4, f5)
+    set_component!(f, k, f1, f2, f3, f4, f5, equations)
   end
 
-  return SVector{nvariables(equations), real(equations)}(f)
+  return SVector(f)
 end
 
 """
@@ -171,7 +170,7 @@ function source_terms_standard(u, x, t, equations::IdealMhdMultiIonEquations1D)
   B1, B2, B3, _ = u
   v1_plus, v2_plus, v3_plus, vk1_plus, vk2_plus, vk3_plus = charge_averaged_velocities(u, equations)
 
-  s = (zero(u[1]), zero(u[1]), zero(u[1]))
+  s = zero(MVector{nvariables(equations), eltype(u)})
   for k in eachcomponent(equations)
     rho, rho_v1, rho_v2, rho_v3, rho_e = get_component(k, u, equations)
     v1 = rho_v1 / rho
@@ -186,10 +185,10 @@ function source_terms_standard(u, x, t, equations::IdealMhdMultiIonEquations1D)
     s4 = r_rho * (v1_diff * B2 - v2_diff - B1)
     s5 = v1 * s2 + v2 * s3 + v3 * s4
 
-    s = (s..., zero(u[1]), s2, s3, s4, s5)
+    set_component!(s, k, 0, s2, s3, s4, s5, equations)
   end
 
-  return SVector{nvariables(equations), real(equations)}(s)
+  return SVector(s)
 end
 
 """
@@ -215,7 +214,7 @@ The term is composed of three parts
   mag_norm_avg = 0.5*(mag_norm_ll+mag_norm_rr)
 
   # Compute charge ratio of u_ll
-  charge_ratio_ll = zeros(typeof(u_ll[1]), ncomponents(equations))
+  charge_ratio_ll = zero(MVector{ncomponents(equations), eltype(u_ll)})
   total_electron_charge = zero(u_ll[1])
   for k in eachcomponent(equations)
     rho_k = u_ll[(k-1)*5+4]
@@ -228,7 +227,7 @@ The term is composed of three parts
   v1_plus_ll, v2_plus_ll, v3_plus_ll, vk1_plus_ll, vk2_plus_ll, vk3_plus_ll = charge_averaged_velocities(u_ll, equations)
   v1_plus_rr, v2_plus_rr, v3_plus_rr, vk1_plus_rr, vk2_plus_rr, vk3_plus_rr = charge_averaged_velocities(u_rr, equations)
   
-  f = (zero(u_ll[1]), zero(u_ll[1]), zero(u_ll[1]))
+  f = zero(MVector{nvariables(equations), eltype(u_ll)})
   # TODO: Add entries of Powell term for induction equation
   for k in eachcomponent(equations)
     # Compute Powell (only needed for non-constant B1)
@@ -260,10 +259,10 @@ The term is composed of three parts
     f5 = 2 * f5 - B2_ll * (vk1_minus_ll * B2_ll - vk2_minus_ll * B1_ll) - B3_ll * (vk1_minus_ll * B3_ll - vk3_minus_ll * B1_ll)
     
     # Append to the flux vector
-    f = (f..., zero(u_ll[1]), f2, f3, f4, f5)
+    set_component!(f, k, 0, f2, f3, f4, f5, equations)
   end
 
-  return SVector{nvariables(equations), real(equations)}(f)
+  return SVector(f)
 end
 
 """
@@ -283,7 +282,7 @@ The term is composed of three parts
   mag_norm_rr = B1_rr^2 + B2_rr^2 + B3_rr^2
 
   # Compute charge ratio of u_ll
-  charge_ratio_ll = zeros(typeof(u_ll[1]), ncomponents(equations))
+  charge_ratio_ll = zero(MVector{ncomponents(equations), eltype(u_ll)})
   total_electron_charge = zero(u_ll[1])
   for k in eachcomponent(equations)
     rho_k = u_ll[(k-1)*5+4]
@@ -295,7 +294,7 @@ The term is composed of three parts
   # Compute auxiliary variables
   v1_plus_rr, v2_plus_rr, v3_plus_rr, vk1_plus_rr, vk2_plus_rr, vk3_plus_rr = charge_averaged_velocities(u_rr, equations)
   
-  f = (zero(u_ll[1]), zero(u_ll[1]), zero(u_ll[1]))
+  f = zero(MVector{nvariables(equations), eltype(u_ll)})
   # TODO: Add entries of Powell term for induction equation
   for k in eachcomponent(equations)
     # Compute Powell (only needed for non-constant B1)
@@ -318,10 +317,10 @@ The term is composed of three parts
     # It's not needed to adjust to Trixi's non-conservative form
 
     # Append to the flux vector
-    f = (f..., zero(u_ll[1]), f2, f3, f4, f5)
+    set_component!(f, k, 0, f2, f3, f4, f5, equations)
   end
 
-  return SVector{nvariables(equations), real(equations)}(f)
+  return SVector(f)
 end
 
 """
@@ -355,13 +354,17 @@ function flux_ruedaramirez_etal(u_ll, u_rr, orientation::Integer, equations::Ide
   mag_norm_rr = B1_rr^2 + B2_rr^2 + B3_rr^2
   mag_norm_avg = 0.5*(mag_norm_ll+mag_norm_rr)
 
+  f = zero(MVector{nvariables(equations), eltype(u_ll)})
+
   # Magnetic field components from f^MHD
   f6 = zero(u_ll[1])
   f7 = v1_plus_avg * B2_avg - v2_plus_avg * B1_avg
   f8 = v1_plus_avg * B3_avg - v3_plus_avg * B1_avg
 
-  # Start concatenating the flux
-  f = (f6, f7, f8)
+  # Start building the flux
+  f[1] = f6
+  f[2] = f7
+  f[3] = f8
 
   # Iterate over all components
   for k in eachcomponent(equations)
@@ -426,10 +429,10 @@ function flux_ruedaramirez_etal(u_ll, u_rr, orientation::Integer, equations::Ide
            + 0.5 * vk1_plus_avg * mag_norm_avg - vk1_plus_avg * B1_avg * B1_avg - vk2_plus_avg * B1_avg * B2_avg - vk3_plus_avg * B1_avg * B3_avg   # Additional terms coming from the MHD non-conservative term (momentum eqs)
            - B2_avg *  (vk1_minus_avg * B2_avg - vk2_minus_avg * B1_avg) - B3_avg * (vk1_minus_avg * B3_avg - vk3_minus_avg * B1_avg) )             # Terms coming from the non-conservative term 3 (induction equation!)
 
-    f = (f..., f1, f2, f3, f4, f5)
+    set_component!(f, k, f1, f2, f3, f4, f5, equations)
   end
 
-  return SVector{nvariables(equations), real(equations)}(f)
+  return SVector(f)
 end
 
 """
@@ -478,7 +481,10 @@ function cons2prim(u, equations::IdealMhdMultiIonEquations1D)
   @unpack gammas = equations
   B1, B2, B3, _ = u
 
-  prim = (B1, B2, B3)
+  prim = zero(MVector{nvariables(equations), eltype(u)})
+  prim[1] = B1
+  prim[2] = B2
+  prim[3] = B3
   for k in eachcomponent(equations)
     rho, rho_v1, rho_v2, rho_v3, rho_e = get_component(k, u, equations)
     srho = 1 / rho
@@ -488,10 +494,11 @@ function cons2prim(u, equations::IdealMhdMultiIonEquations1D)
 
     p = (gammas[k] - 1) * (rho_e - 0.5 * (rho_v1 * v1 + rho_v2 * v2 + rho_v3 * v3
                                  + B1 * B1 + B2 * B2 + B3 * B3))
-    prim = (prim..., rho, v1, v2, v3, p)
+    
+    set_component!(prim, k, rho, v1, v2, v3, p, equations)
   end
 
-  return SVector{nvariables(equations), real(equations)}(prim)
+  return SVector(prim)
 end
 
 """
@@ -502,8 +509,8 @@ Convert conservative variables to entropy
   B1, B2, B3, _ = u
 
   prim = cons2prim(u, equations)
-  entropy = ()
-  rho_p_plus = zero(u[1])
+  entropy = zero(MVector{nvariables(equations), eltype(u)})
+  rho_p_plus = zero(real(equations))
   for k in eachcomponent(equations)
     rho, v1, v2, v3, p = get_component(k, prim, equations)
     s = log(p) - gammas[k] * log(rho)
@@ -514,16 +521,16 @@ Convert conservative variables to entropy
     w4 = rho_p * v3
     w5 = -rho_p
     rho_p_plus += rho_p
-    entropy = (entropy..., w1, w2, w3, w4, w5)
+    
+    set_component!(entropy, k, w1, w2, w3, w4, w5, equations)
   end
 
   # Additional non-conservative variables
-  w6 = rho_p_plus * B1
-  w7 = rho_p_plus * B2
-  w8 = rho_p_plus * B3
-  entropy = (w6, w7, w8, entropy...)
+  entropy[1] = rho_p_plus * B1
+  entropy[2] = rho_p_plus * B2
+  entropy[3] = rho_p_plus * B3
   
-  return SVector{nvariables(equations), real(equations)}(entropy)
+  return SVector(entropy)
 end
 
 
@@ -534,7 +541,10 @@ Convert primitive to conservative variables
   @unpack gammas = equations
   B1, B2, B3, _ = prim
 
-  cons = (B1, B2, B3)
+  cons = zero(MVector{nvariables(equations), eltype(prim)})
+  cons[1] = B1
+  cons[2] = B2
+  cons[3] = B3
   for k in eachcomponent(equations)
     rho, v1, v2, v3, p = get_component(k, prim, equations)
     rho_v1 = rho * v1
@@ -543,7 +553,8 @@ Convert primitive to conservative variables
 
     rho_e = p/(gammas[k] - 1.0) + 0.5 * (rho_v1 * v1 + rho_v2 * v2 + rho_v3 * v3) +
                                   0.5 * (B1^2 + B2^2 + B3^2)
-    cons = (cons..., rho, rho_v1, rho_v2, rho_v3, rho_e)
+    
+    set_component!(cons, k, rho, rho_v1, rho_v2, rho_v3, rho_e, equations)
   end
 
   return SVector{nvariables(equations), real(equations)}(cons)
@@ -589,9 +600,9 @@ Routine to compute the charge-averaged velocities:
 
   total_electron_charge = zero(u[1])
   
-  vk1_plus = zeros(typeof(u[1]), ncomponents(equations))
-  vk2_plus = zeros(typeof(u[1]), ncomponents(equations))
-  vk3_plus = zeros(typeof(u[1]), ncomponents(equations))
+  vk1_plus = zero(MVector{ncomponents(equations), eltype(u)})
+  vk2_plus = zero(MVector{ncomponents(equations), eltype(u)})
+  vk3_plus = zero(MVector{ncomponents(equations), eltype(u)})
 
   for k in eachcomponent(equations)
     rho_k = u[(k-1)*5+4]
@@ -610,16 +621,31 @@ Routine to compute the charge-averaged velocities:
   v2_plus = sum(vk2_plus)
   v3_plus = sum(vk3_plus)
 
-  return v1_plus, v2_plus, v3_plus, SVector{ncomponents(equations), real(equations)}(vk1_plus),
-                                    SVector{ncomponents(equations), real(equations)}(vk2_plus),
-                                    SVector{ncomponents(equations), real(equations)}(vk3_plus)
+  return v1_plus, v2_plus, v3_plus, SVector(vk1_plus), SVector(vk2_plus), SVector(vk3_plus)
 end
 
 """
 Get the flow variables of component k
 """
 @inline function get_component(k, u, equations::IdealMhdMultiIonEquations1D)
-  return SVector{5, real(equations)}(u[(k-1)*5+4:(k-1)*5+8])
+  # The first 3 entries of u contain the magnetic field. The following entries contain the density, momentum (3 entries), and energy of each component.
+  return SVector(u[3 + (k - 1) * 5 + 1],
+                 u[3 + (k - 1) * 5 + 2],
+                 u[3 + (k - 1) * 5 + 3],
+                 u[3 + (k - 1) * 5 + 4],
+                 u[3 + (k - 1) * 5 + 5])
+end
+
+"""
+Set the flow variables of component k
+"""
+@inline function set_component!(u, k, u1, u2, u3, u4, u5, equations::IdealMhdMultiIonEquations1D)
+  # The first 3 entries of u contain the magnetic field. The following entries contain the density, momentum (3 entries), and energy of each component.
+  u[3 + (k - 1) * 5 + 1] = u1
+  u[3 + (k - 1) * 5 + 2] = u2
+  u[3 + (k - 1) * 5 + 3] = u3
+  u[3 + (k - 1) * 5 + 4] = u4
+  u[3 + (k - 1) * 5 + 5] = u5
 end
 
 @inline function density_product(u, equations::IdealMhdMultiIonEquations1D)