For `trixi_include(joinpath(examples_dir(), "tree_2d_dgsem", "elixir_mhdmultiion_ec.jl"), tspan=(0.0, 0.01))`, I get approximately a speedup of 100x for the `time/DOF/rhs!` metric. There were two main causes for the spurious allocations:
* Type instabilities due to an extension of a tuple with splatting:
  ```julia
  f = (a, b, c)
  [...]
  f = (f..., d, e, f)
  ```
  where the extension was in a loop. This caused `f` to be inferred as `Tuple{Vararg{Float64}}`, i.e., Julia was not able to figure out how long this tuple would be at the end of the loop.
* Use of `zeros(Float64, some_length)`, i.e., explicit heap allocations
* Memory allocations in `get_component` due to array slicing (which causes the data to be copied)

For me, the easiest approach (in terms of implementation, but also conceptually) was to use `MVector`s, which are mutable versions of `SVector` and which, according to @ranocha, do not cause allocations if they do not leave a function scope. This non-allocating behavior was confirmed in my tests.

What's missing is some allocations in `calc_sources!`, which I was not able to further track down, since they did not occur when I manually called the function using `@btime` 🤷‍♂️ Maybe @ranocha has an idea to proceed further?

For future reference, @amrueda, at the bottom of this post I have copied the auxiliary code I used to set up a simulation such that I could benchmark individual functions.

### Before (branch `multi_ion_mhd`)
`time/DOF/rhs!: 5.52488354e-05`

**Timer output:**
```
 ────────────────────────────────────────────────────────────────────────────────────
              Trixi.jl                      Time                    Allocations
                                   ───────────────────────   ────────────────────────
         Tot / % measured:              4.25s / 100.0%           3.17GiB / 100.0%

 Section                   ncalls     time    %tot     avg     alloc    %tot      avg
 ────────────────────────────────────────────────────────────────────────────────────
 rhs!                          16    3.59s   84.6%   224ms   2.70GiB   85.2%   173MiB
   volume integral             16    2.88s   67.8%   180ms   2.28GiB   71.9%   146MiB
   source terms                16    453ms   10.7%  28.3ms    258MiB    7.9%  16.1MiB
   interface flux              16    255ms    6.0%  15.9ms    173MiB    5.3%  10.8MiB
   prolong2interfaces          16   2.14ms    0.1%   133μs     0.00B    0.0%    0.00B
   surface integral            16    979μs    0.0%  61.2μs     0.00B    0.0%    0.00B
   ~rhs!~                      16    138μs    0.0%  8.61μs   9.33KiB    0.0%     597B
   reset ∂u/∂t                 16    131μs    0.0%  8.21μs     0.00B    0.0%    0.00B
   Jacobian                    16    126μs    0.0%  7.85μs     0.00B    0.0%    0.00B
   prolong2mortars             16   12.8μs    0.0%   799ns     0.00B    0.0%    0.00B
   prolong2boundaries          16   10.8μs    0.0%   676ns     0.00B    0.0%    0.00B
   mortar flux                 16   7.74μs    0.0%   484ns     0.00B    0.0%    0.00B
   boundary flux               16   1.46μs    0.0%  91.2ns     0.00B    0.0%    0.00B
 analyze solution               2    634ms   14.9%   317ms    462MiB   14.2%   231MiB
 I/O                            3   15.9ms    0.4%  5.29ms   10.1MiB    0.3%  3.36MiB
   save solution                2   14.3ms    0.3%  7.15ms   10.0MiB    0.3%  5.02MiB
   ~I/O~                        3   1.54ms    0.0%   512μs   34.1KiB    0.0%  11.4KiB
   get element variables        2   18.0μs    0.0%  9.01μs   3.97KiB    0.0%  1.98KiB
   save mesh                    2    240ns    0.0%   120ns     0.00B    0.0%    0.00B
 calculate dt                   4   4.41ms    0.1%  1.10ms   9.00MiB    0.3%  2.25MiB
 ────────────────────────────────────────────────────────────────────────────────────
```

### After (this PR):
`time/DOF/rhs!: 5.80463853e-07`

**Timer output:**
```
 ────────────────────────────────────────────────────────────────────────────────────
              Trixi.jl                      Time                    Allocations
                                   ───────────────────────   ────────────────────────
         Tot / % measured:              252ms /  99.7%            118MiB /  99.6%

 Section                   ncalls     time    %tot     avg     alloc    %tot      avg
 ────────────────────────────────────────────────────────────────────────────────────
 analyze solution               2    196ms   78.1%  98.0ms    109MiB   92.7%  54.4MiB
 rhs!                          16   38.0ms   15.1%  2.37ms   14.6KiB    0.0%     933B
   volume integral             16   29.8ms   11.9%  1.87ms     0.00B    0.0%    0.00B
   interface flux              16   5.16ms    2.1%   322μs     0.00B    0.0%    0.00B
   prolong2interfaces          16   1.12ms    0.4%  70.2μs     0.00B    0.0%    0.00B
   source terms                16   1.06ms    0.4%  66.5μs   5.25KiB    0.0%     336B
   surface integral            16    525μs    0.2%  32.8μs     0.00B    0.0%    0.00B
   Jacobian                    16    117μs    0.0%  7.34μs     0.00B    0.0%    0.00B
   reset ∂u/∂t                 16    114μs    0.0%  7.14μs     0.00B    0.0%    0.00B
   ~rhs!~                      16   25.1μs    0.0%  1.57μs   9.33KiB    0.0%     597B
   prolong2mortars             16   1.37μs    0.0%  85.6ns     0.00B    0.0%    0.00B
   prolong2boundaries          16   1.20μs    0.0%  75.0ns     0.00B    0.0%    0.00B
   mortar flux                 16    880ns    0.0%  55.0ns     0.00B    0.0%    0.00B
   boundary flux               16    340ns    0.0%  21.2ns     0.00B    0.0%    0.00B
 I/O                            3   15.9ms    6.4%  5.31ms   8.57MiB    7.3%  2.86MiB
   save solution                2   13.8ms    5.5%  6.91ms   8.54MiB    7.3%  4.27MiB
   ~I/O~                        3   2.10ms    0.8%   700μs   34.1KiB    0.0%  11.4KiB
   get element variables        2   14.4μs    0.0%  7.21μs   3.97KiB    0.0%  1.98KiB
   save mesh                    2    170ns    0.0%  85.0ns     0.00B    0.0%    0.00B
 calculate dt                   4   1.11ms    0.4%   278μs     0.00B    0.0%    0.00B
 ────────────────────────────────────────────────────────────────────────────────────
```

This is the file I used as a starting point for benchmarking (`mwe.jl`); the function that I benchmarked of course changed later on:
```julia
using Trixi, BenchmarkTools

@info "Running elixir..."
trixi_include(joinpath(examples_dir(), "tree_2d_dgsem", "elixir_mhdmultiion_ec.jl"), tspan=(0.0, 0.01))

@info "Creating dummy data..."
u_ode = deepcopy(ode.u0)
du_ode = similar(u_ode)

u = Trixi.wrap_array(u_ode, semi)
du = Trixi.wrap_array(du_ode, semi)
u_node = Trixi.get_node_vars(u, equations, solver, 1, 1, 1)

@info "Running rhs!..."
@time Trixi.rhs!(du, u, 0.0, mesh, equations, semi.initial_condition, semi.boundary_conditions,
                 semi.source_terms, solver, semi.cache)

@info "Benchmarking charge_averaged_velocities..."
@btime Trixi.charge_averaged_velocities($u_node, $equations)
```
Note that not all code is strictly necessary, e.g., you do not necessarily need to set up all these data structures to be able to call `rhs!`, but it helped me to descend gently into the madness 😬