@ranocha 2023-05-06
I guess we pass a function to the function calculating the source terms without explicit specialization.