@sloede 2023-05-06
> I guess we pass a function to the function calculating the source terms without explicit specialization.

You mean this might be a case of https://docs.julialang.org/en/v1/manual/performance-tips/#Be-aware-of-when-Julia-avoids-specializing? In that case, we pay only a small allocation per rhs invocation, which I guess is acceptable.