@ranocha 2023-05-06
From your initial post, you can see that it's only 336 Bytes per RHS call for the source terms. That looks like one dynamic dispatch to me