@sloede 2023-05-08
> I would prefer if we could continue the discussion in the original PR - not a PR to a PR. This makes it easier to review for me. You can of course resolve all open discussions here, but please consider merging this PR sooner rather than later, without too many additional changes on top of the basic performance improvements.

Completely agree... all discussions here will be hard to find once this is merged. @amrueda I'd say just merge now and then figure out the rest later